backend/requirements.txt
# NOF1 Trading Bot - Python Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Crypto/Trading
python-binance==1.0.19

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0

# Utilities
python-dateutil==2.8.2

# WebSocket Support
websockets==12.0

# Logging
loguru==0.7.2

